the generated region sizes add up to the input number with sometimes 1 unit of difference #5

+++ src/MathOp.h
@@ -57,7 +57,7 @@ public:
 
     static double frobenius_avg(vector<vector<int>>& mat, vector<vector<int>>& ground_truth);
 
-    static vector<double> dirichlet_sample(int len, double alpha=1.0);
+    static vector<long double> dirichlet_sample(int len, double alpha = 1.0);
 
 
 };
